refactor: modularize transformQuiverNoteToObsidian function in preparation for upcoming feature enhancements
